+++ packages/sampler/stories/moonstone-stories/VirtualGridList.js
@@ -15,7 +15,6 @@ const
 	prop = {
 		direction: {'horizontal': 'horizontal', 'vertical': 'vertical'}
 	},
-	listStyle = {height: ri.scale(550) + 'px'},
 	items = [],
 	// eslint-disable-next-line enact/prop-types
 	renderItem = ({data, index, ...rest}) => {
@@ -53,11 +52,16 @@ storiesOf('VirtualList.VirtualGridList')
 				dataSize={number('dataSize', items.length)}
 				direction={select('direction', prop.direction, 'vertical')}
 				focusableScrollbar={nullify(boolean('focusableScrollbar', false))}
-				itemSize={{minWidth: ri.scale(number('minWidth', 180)), minHeight: ri.scale(number('minHeight', 270))}}
+				itemSize={{
+					minWidth: ri.scale(number('minWidth', 180)),
+					minHeight: ri.scale(number('minHeight', 270))
+				}}
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
 				spacing={ri.scale(number('spacing', 20))}
-				style={listStyle}
+				style={{
+					height: ri.unit(549, 'rem')
+				}}
 			/>
 		),
 		{propTables: [Config]}