+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,12 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/DatePicker` to calculate min and max year in the current calender
+## [1.13.3] - 2017-01-16
+
+### Fixed
+
+- `moonstone/TimePicker` to not read out meridiem label when meridiem picker gets a focus
+- `moonstone/Scroller` to correctly update scrollbars when the scroller's contents change
 
 ## [1.13.2] - 2017-12-14